# api/openapi.yaml
openapi: 3.0.3
info:
  title: Saviynt External Connector Framework API Specification
  description: "This page lists the API specification for Saviynt External Connector\
    \ Framework. You can use \"Try it out\" option to explore the API specification.\
    \ \"Try it out\" will require you to authorize with a static token, so you should\
    \ first click on Authorize button at the top and provide a value, this is a sample\
    \ specification and there is no actual token validation, so you can provide any\
    \ string value for e.g. MyBearerToken and click on Authorize, then Close."
  version: "1.0"
servers:
- url: "{server}/api/{version}"
  variables:
    server:
      default: https://example.com:443
    version:
      default: v1
security:
- bearerAuth: []
...
paths:
...
  /createAccount:
    post:
      tags:
      - Account Provisioning
      summary: This API creates an account
      operationId: postCreateAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createaccount'
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
        "400":
          description: Failure in create account
      x-swagger-router-controller: AccountProvisioning
  /updateAccount/{accountID}:
    patch:
      tags:
      - Account Provisioning
      summary: This API updates an account
      operationId: patchUpdateAccount
      parameters:
      - name: accountID
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
            examples:
              Update Account:
                value:
                  firstname: Logan
                  preferedFirstName: Logan
                  lastname: Prasad
                  middlename: "N"
                  email: logan@example.net
                  employeeType: Regular
                  jobcodedesc: QA
                  manager: Joe
                  statuskey: "1"
                  customproperty1: customized value
              Enable Account:
                value:
                  status: "1"
              Disable Account:
                value:
                  status: "0"
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              examples:
                Update Account:
                  value:
                    errorCode: "0"
                    message: SUCCESS Account Updated Successfully
                Enable Account:
                  value:
                    errorCode: "0"
                    message: SUCCESS Account Enabled Successfully
                Disable Account:
                  value:
                    errorCode: "0"
                    message: SUCCESS Account Disabled Successfully
        "400":
          description: Failure response
      x-swagger-router-controller: AccountProvisioning
  /removeAccount/{accountID}:
    delete:
      tags:
      - Account Provisioning
      summary: This API deletes an account
      operationId: deleteRemoveAccount
      parameters:
      - name: accountID
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
        "400":
          description: Failure in account deletion
      x-swagger-router-controller: AccountProvisioning
  /changePassword:
    put:
      tags:
      - Account Provisioning
      summary: This API can be used to change password of an account
      operationId: put_changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/changePassword_body'
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
        "400":
          description: Failure in change password
      x-swagger-router-controller: AccountProvisioning
  /entitlement/group/addAccess:
    put:
      tags:
      - Access Provisioning
      summary: This API grants group access to an account
      operationId: put_entitlementgroups
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/group_addAccess_body'
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_4'
        "400":
          description: Failure response
      x-swagger-router-controller: AccessProvisioning
  /entitlement/role/addAccess:
    put:
      tags:
      - Access Provisioning
      summary: This API grants role access to an account
      operationId: entitlementRoleAddAccessPUT
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/role_addAccess_body'
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_5'
        "400":
          description: Failure response
      x-swagger-router-controller: AccessProvisioning
  /entitlement/group/removeAccess:
    put:
      tags:
      - Access Provisioning
      summary: This API removes group access from an account
      operationId: entitlementGroupRemoveAccessPUT
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/group_removeAccess_body'
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_6'
        "400":
          description: Failure response
      x-swagger-router-controller: AccessProvisioning
  /entitlement/role/removeAccess:
    put:
      tags:
      - Access Provisioning
      summary: This API removes role access from an account
      operationId: entitlementRoleRemoveAccessPUT
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/role_removeAccess_body'
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_7'
        "400":
          description: Failure response
      x-swagger-router-controller: AccessProvisioning
components:
  schemas:
...
    createaccount:
      type: object
      properties:
        accountID:
          type: string
          example: "15553759289490"
        firstname:
          type: string
          example: Logan
        preferedFirstName:
          type: string
          example: Logan
        lastname:
          type: string
          example: Prasad
        middlename:
          type: string
          example: "N"
        email:
          type: string
          example: logan@example.net
        employeeType:
          type: string
          example: Regular
        jobcodedesc:
          type: string
          example: QA
        manager:
          type: string
          example: Joe
        statuskey:
          type: string
          example: "1"
        customproperty1:
          type: string
          example: customized value
...
    inline_response_200_1:
      type: object
      properties:
        errorCode:
          type: string
          example: "0"
        message:
          type: string
          example: SUCCESS Account Created Successfully
      example:
        errorCode: "0"
        message: SUCCESS Account Created Successfully
    inline_response_200_2:
      type: object
      properties:
        errorCode:
          type: string
          example: "0"
        message:
          type: string
          example: SUCCESS Account Removed Successfully
      example:
        errorCode: "0"
        message: SUCCESS Account Removed Successfully
    changePassword_body:
      type: object
      properties:
        accountID:
          type: string
          example: "15553759289492"
        password:
          type: string
          example: EnterPasswordHere
    inline_response_200_3:
      type: object
      properties:
        errorCode:
          type: string
          example: "0"
        message:
          type: string
          example: SUCCESS Account Password Updated Successfully
      example:
        errorCode: "0"
        message: SUCCESS Account Password Updated Successfully
    group_addAccess_body:
      type: object
      properties:
        accountID:
          type: string
          example: "15553759289492"
        entitlementID:
          type: string
          example: "15553798176145"
    inline_response_200_4:
      type: object
      properties:
        errorCode:
          type: string
          example: "0"
        message:
          type: string
          example: SUCCESS Group Access Granted Successfully
      example:
        errorCode: "0"
        message: SUCCESS Group Access Granted Successfully
    role_addAccess_body:
      type: object
      properties:
        accountID:
          type: string
          example: "15553759289492"
        entitlementID:
          type: string
          example: "15553670302865"
    inline_response_200_5:
      type: object
      properties:
        errorCode:
          type: string
          example: "0"
        message:
          type: string
          example: SUCCESS Role Access Granted Successfully
      example:
        errorCode: "0"
        message: SUCCESS Role Access Granted Successfully
    group_removeAccess_body:
      type: object
      properties:
        accountID:
          type: string
          example: "15553759289492"
        entitlementID:
          type: string
          example: "15553798176145"
    inline_response_200_6:
      type: object
      properties:
        errorCode:
          type: string
          example: "0"
        message:
          type: string
          example: SUCCESS Group Access Revoked Successfully
      example:
        errorCode: "0"
        message: SUCCESS Group Access Revoked Successfully
    role_removeAccess_body:
      type: object
      properties:
        accountID:
          type: string
          example: "15553759289492"
        entitlementID:
          type: string
          example: "15553670302865"
    inline_response_200_7:
      type: object
      properties:
        errorCode:
          type: string
          example: "0"
        message:
          type: string
          example: SUCCESS Role Access Removed Successfully
      example:
        errorCode: "0"
        message: SUCCESS Role Access Removed Successfully
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT